Move OnboardingScreen out of OnboardingActivity
